+++ model/FloatingPageTurnAction.ts
@@ -1,0 +1,6 @@
+export enum FloatingPageTurnAction {
+  CenterViewKeepZoom = "CenterViewKeepZoom",
+  CenterViewResetZoom = "CenterViewResetZoom",
+  FitToScreen = "FitToScreen",
+  KeepCurrentView = "KeepCurrentView",
+}