+++ model/PageView.ts
@@ -1,0 +1,4 @@
+export enum PageView {
+  Fixed = "Fixed",
+  Floating = "Floating",
+}